% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_charts.R
\name{display_charts}
\alias{display_charts}
\title{Display Charts Interactively}
\usage{
display_charts(filename)
}
\arguments{
\item{filename}{A CSV file containing the data}
}
\description{
A Shiny application for interactive exploration of the Questionnaire Data
}
\note{
This function is a wrapper for \code{\link[shiny]{runApp}} and also
provides for interactive selection of the CSV file.
}
