% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data.R
\name{display_data}
\alias{display_data}
\title{Display Charts Interactively}
\usage{
display_data(filename)
}
\arguments{
\item{filename}{A CSV file containing the data}
}
\description{
A Shiny application for interactive exploration of the Questionnaire Data
}
\note{
This function is a wrapper for \code{\link[shiny]{runApp}} and also
provides for interactive selection of the CSV file.
}
