% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_barchart_all}
\alias{show_barchart_all}
\title{Display the categorical data using bar charts
data on open-ended questions are removed from the data}
\usage{
show_barchart_all(file = NULL, data = NULL)
}
\arguments{
\item{file}{A path to text file containing data in CSV format}

\item{data}{A R object of class \code{data.frame} containing questionnaire 
data}
}
\value{
The function does not return a value but displays bar charts
}
\description{
Display the categorical data using bar charts
data on open-ended questions are removed from the data
}
\note{
This function is designed for a specific data set, not for general use.
}
