% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data.R
\name{display_data}
\alias{display_data}
\title{Display Data Interactively}
\usage{
display_data(filename)
}
\arguments{
\item{filename}{A comma-separated values (CSV) file containing the data.}
}
\description{
A \emph{Shiny} application for the interactive display and exploration of
the questionnaire data.
}
\note{
This function is a wrapper for \code{\link[shiny]{runApp}}. Also, the
function works only with a dataset that has a specific set of variables.
}
